+++ src/System.Runtime/tests/EnumTests.cs
