+++ src/System.Runtime/tests/System.Runtime.Tests.csproj
@@ -148,40 +148,31 @@
     <Compile Include="$(CommonTestPath)\System\Collections\IEnumerable.Generic.Serialization.Tests.cs">
       <Link>Common\System\Collections\IEnumerable.Generic.Serialization.Tests.cs</Link>
     </Compile>
-    <Compile Include="ArrayTests.netstandard.cs" />
-    <Compile Include="System\DelegateTests.netstandard.cs" />
-    <Compile Include="System\AttributeTests.netstandard.cs" />
+    <Compile Include="ArrayTests.cs" />
     <Compile Include="CompareToTests.cs" />
-    <Compile Include="DBNullTests.netstandard.cs" />
-    <Compile Include="DecimalTests.netstandard.cs" />
-    <Compile Include="EnumTests.netstandard.cs" />
+    <Compile Include="DBNullTests.cs" />
+    <Compile Include="DecimalTests.cs" />
+    <Compile Include="EnumTests.cs" />
     <Compile Include="GetTypeCodeTests.cs" />
     <Compile Include="System\AccessViolationExceptionTests.cs" />
     <Compile Include="System\BadImageFormatExceptionTests.cs" />
-    <Compile Include="System\CharTests.netstandard.cs" />
-    <Compile Include="System\DateTimeOffsetTests.netstandard.cs" />
     <Compile Include="System\DuplicateWaitObjectExceptionTests.cs" />
     <Compile Include="System\ExternalExceptionTests.cs" />
     <Compile Include="System\ApplicationExceptionTests.cs" />
     <Compile Include="System\EntryPointNotFoundExceptionTests.cs" />
     <Compile Include="System\ExceptionTests.cs" />
     <Compile Include="System\ExecutionEngineExceptionTests.cs" />
-    <Compile Include="System\GuidTests.netstandard.cs" />
     <Compile Include="System\MarshalByRefObjectTests.cs" />
     <Compile Include="System\MissingMemberExceptionTests.cs" />
     <Compile Include="System\NotFiniteNumberExceptionTests.cs" />
     <Compile Include="System\Security\SecurityAttributeTests.cs" />
     <Compile Include="System\StackOverflowExceptionTests.cs" />
     <Compile Include="System\SystemExceptionTests.cs" />
-    <Compile Include="System\StringTests.netstandard.cs" />
-    <Compile Include="System\TimeSpanTests.netstandard.cs" />
     <Compile Include="System\TypeUnloadedExceptionTests.cs" />
-    <Compile Include="TimeZoneTests.netstandard.cs" />
-    <Compile Include="TimeZoneInfoTests.netstandard.cs" />
-    <Compile Include="TimeZoneNotFoundExceptionTests.netstandard.cs" />
-    <Compile Include="VersionTests.netstandard.cs" />
-    <Compile Include="System\ComponentModel\DefaultValueAttributeTests.netstandard.cs" />
-    <Compile Include="System\ComponentModel\EditorBrowsableAttributeTests.netstandard.cs" />
+    <Compile Include="TimeZoneTests.cs" />
+    <Compile Include="TimeZoneInfoTests.cs" />
+    <Compile Include="TimeZoneNotFoundExceptionTests.cs" />
+    <Compile Include="VersionTests.cs" />
     <Compile Include="System\Reflection\AssemblyTests.cs" />
     <Compile Include="System\Reflection\AssemblyNameTests.cs" />
     <Compile Include="System\Reflection\ConstructorInfoTests.cs" />
@@ -205,9 +196,8 @@
     <EmbeddedResource Include="System\Reflection\EmbeddedTextFile.txt">
       <LogicalName>System.Reflection.Tests.EmbeddedTextFile.txt</LogicalName>
     </EmbeddedResource>
-    <Compile Include="System\Threading\WaitHandleTests.netstandard.cs" />
-    <Compile Include="Microsoft\Win32\SafeHandles\CriticalHandleZeroOrMinusOneIsInvalid.netstandard.cs" />
-    <Compile Include="Microsoft\Win32\SafeHandles\SafeHandleZeroOrMinusOneIsInvalid.netstandard.cs" />
+    <Compile Include="Microsoft\Win32\SafeHandles\CriticalHandleZeroOrMinusOneIsInvalid.cs" />
+    <Compile Include="Microsoft\Win32\SafeHandles\SafeHandleZeroOrMinusOneIsInvalid.cs" />
     <Compile Include="System\Reflection\ModuleTests.cs" />
     <Compile Include="System\Reflection\PointerTests.cs" />
     <Compile Include="System\Reflection\RuntimeReflectionExtensionsTests.cs" />
@@ -216,18 +206,13 @@
     <Compile Include="$(CommonTestPath)\System\Runtime\Serialization\Formatters\BinaryFormatterHelpers.cs">
       <Link>System\Runtime\Serialization\Formatters\BinaryFormatterHelpers.cs</Link>
     </Compile>
-    <Compile Include="System\Runtime\CompilerServices\RuntimeHelpersTests.netstandard.cs" />
     <Compile Include="System\Runtime\NgenServicingAttributesTests.cs" />
-    <Compile Include="System\Attributes.netstandard.cs" />
-    <Compile Include="System\HandleTests.netstandard.cs" />
+    <Compile Include="System\Attributes.cs" />
+    <Compile Include="System\HandleTests.cs" />
     <Compile Include="System\AppContext\AppContext.Switch.cs" />
     <Compile Include="System\AppContext\AppContext.Switch.Validation.cs" />
-    <Compile Include="System\Reflection\StrongNameKeyPairTests.netstandard.cs" />
-    <Compile Include="System\ActivatorTest.netstandard.cs" />
-    <Compile Include="System\ExceptionTests.netstandard.cs" />
-    <Compile Include="System\TypeTests.netstandard.cs" />
+    <Compile Include="System\Reflection\StrongNameKeyPairTests.cs" />
     <Compile Include="System\Runtime\Tests.netstandard.1.7.cs" />
-    <Compile Include="System\GCTests.netstandard.cs" />
     <Compile Include="$(CommonTestPath)\System\Diagnostics\RemoteExecutorTestBase.cs">
       <Link>Common\System\Diagnostics\RemoteExecutorTestBase.cs</Link>
     </Compile>
