--- src/System.Runtime/tests/System/ActivatorTest.netstandard.cs
@@ -1,121 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-using System;
-using System.Globalization;
-using System.Reflection;
-using System.Runtime.InteropServices;
-using Xunit;
-
-namespace System.Tests
-{
-
-    public class ActivatorTestsExtended
-    {
-        class ClassWithPrivateCtor
-        {
-            static ClassWithPrivateCtor() { Flag.Reset(100); }
-            private ClassWithPrivateCtor() { Flag.Increase(200); }
-            public ClassWithPrivateCtor(int i) { Flag.Increase(300); }
-        }
-        class ClassWithPrivateCtor2
-        {
-            static ClassWithPrivateCtor2() { Flag.Reset(100); }
-            private ClassWithPrivateCtor2() { Flag.Increase(200); }
-            public ClassWithPrivateCtor2(int i) { Flag.Increase(300); }
-        }
-        class ClassWithPrivateCtor3
-        {
-            static ClassWithPrivateCtor3() { Flag.Reset(100); }
-            private ClassWithPrivateCtor3() { Flag.Increase(200); }
-            public ClassWithPrivateCtor3(int i) { Flag.Increase(300); }
-        }
-
-
-        public class IsTestedAttribute : Attribute
-        {
-            private bool flag;
-            public IsTestedAttribute(bool flag)
-            {
-                this.flag = flag;
-
-            }
-        }
-        [IsTestedAttribute(false)]
-        class ClassWithIsTestedAttribute { }
-        [Serializable]
-        class ClassWithSerializableAttribute { }
-        [IsTestedAttribute(false)]
-        class MBRWithIsTestedAttribute : MarshalByRefObject { }
-        class Flag
-        {
-            public static int cnt = 0;
-            public static void Reset(int i) { cnt = i; }
-            public static void Increase(int i) { cnt += i; }
-            public static bool Equal(int i) { return cnt == i; }
-        }
-
-        [Fact]
-        static void TestingBindingFlags()
-        {
-            Assert.Throws<MissingMethodException>(() => Activator.CreateInstance(typeof(ClassWithPrivateCtor), BindingFlags.Public | BindingFlags.Instance, null, null, CultureInfo.CurrentCulture));
-            Assert.Throws<MissingMethodException>(() => Activator.CreateInstance(typeof(ClassWithPrivateCtor), BindingFlags.Public | BindingFlags.Instance, null, new object[] { 1, 2, 3 }, CultureInfo.CurrentCulture));
-
-
-            Flag.Reset(0); Assert.Equal(0, Flag.cnt);
-            Activator.CreateInstance(typeof(ClassWithPrivateCtor), BindingFlags.Static | BindingFlags.NonPublic, null, null, CultureInfo.CurrentCulture);
-            Assert.Equal(300, Flag.cnt);
-            Activator.CreateInstance(typeof(ClassWithPrivateCtor), BindingFlags.Instance | BindingFlags.NonPublic, null, null, CultureInfo.CurrentCulture);
-            Assert.Equal(500, Flag.cnt);
-
-            Flag.Reset(0); Assert.Equal(0, Flag.cnt);
-            Activator.CreateInstance(typeof(ClassWithPrivateCtor2), BindingFlags.Instance | BindingFlags.NonPublic, null, null, CultureInfo.CurrentCulture);
-            Assert.Equal(300, Flag.cnt);
-            Activator.CreateInstance(typeof(ClassWithPrivateCtor2), BindingFlags.Static | BindingFlags.NonPublic, null, null, CultureInfo.CurrentCulture);
-            Assert.Equal(500, Flag.cnt);
-
-            Flag.Reset(0); Assert.Equal(0, Flag.cnt);
-            Activator.CreateInstance(typeof(ClassWithPrivateCtor3), BindingFlags.Instance | BindingFlags.Public, null, new object[] { 122 }, CultureInfo.CurrentCulture);
-            Assert.Equal(400, Flag.cnt);
-            Activator.CreateInstance(typeof(ClassWithPrivateCtor3), BindingFlags.Static | BindingFlags.NonPublic, null, null, CultureInfo.CurrentCulture);
-            Assert.Equal(600, Flag.cnt);
-            Activator.CreateInstance(typeof(ClassWithPrivateCtor3), BindingFlags.Instance | BindingFlags.Public, null, new object[] { 122 }, CultureInfo.CurrentCulture);
-            Assert.Equal(900, Flag.cnt);
-
-            Assert.Throws<MissingMethodException>(() => Activator.CreateInstance(typeof(ClassWithPrivateCtor), BindingFlags.Public | BindingFlags.Static, null, null, CultureInfo.CurrentCulture));
-            Assert.Throws<MissingMethodException>(() => Activator.CreateInstance(typeof(ClassWithPrivateCtor), BindingFlags.Public | BindingFlags.Static, null, new object[] { 122 }, CultureInfo.CurrentCulture));
-        }
-
-        [Fact]
-        static void TestingBindingFlags1()
-        {
-            Assert.Throws<MissingMethodException>(() => Activator.CreateInstance(typeof(ClassWithPrivateCtor), BindingFlags.Public | BindingFlags.Instance, null, null, CultureInfo.CurrentCulture, null));
-            Assert.Throws<MissingMethodException>(() => Activator.CreateInstance(typeof(ClassWithPrivateCtor), BindingFlags.Public | BindingFlags.Instance, null, new object[] { 1, 2, 3 }, CultureInfo.CurrentCulture, null));
-
-            Assert.Throws<MissingMethodException>(() => Activator.CreateInstance(typeof(ClassWithPrivateCtor), BindingFlags.Public | BindingFlags.Static, null, null, CultureInfo.CurrentCulture, null));
-            Assert.Throws<MissingMethodException>(() => Activator.CreateInstance(typeof(ClassWithPrivateCtor), BindingFlags.Public | BindingFlags.Static, null, new object[] { 122 }, CultureInfo.CurrentCulture, null));
-        }
-
-        [Fact]
-        [SkipOnTargetFramework(TargetFrameworkMonikers.NetFramework)]
-        static void TestingActivationAttributes()
-        {
-            Assert.Throws<PlatformNotSupportedException>(() => Activator.CreateInstance(typeof(ClassWithIsTestedAttribute), null, new object[] { new Object() }));
-            Assert.Throws<PlatformNotSupportedException>(() => Activator.CreateInstance(typeof(ClassWithIsTestedAttribute), null, new object[] { new IsTestedAttribute(true) }));
-            Assert.Throws<PlatformNotSupportedException>(() => Activator.CreateInstance(typeof(ClassWithSerializableAttribute), null, new object[] { new ClassWithIsTestedAttribute() }));
-            Assert.Throws<PlatformNotSupportedException>(() => Activator.CreateInstance(typeof(MBRWithIsTestedAttribute), null, new object[] { new IsTestedAttribute(true) }));
-
-            Assert.Throws<PlatformNotSupportedException>(() => Activator.CreateInstance(typeof(ClassWithIsTestedAttribute), 0, null, null, CultureInfo.CurrentCulture, new object[] { new IsTestedAttribute(true) }));
-            Assert.Throws<PlatformNotSupportedException>(() => Activator.CreateInstance(typeof(ClassWithSerializableAttribute), 0, null, null, CultureInfo.CurrentCulture, new object[] { new ClassWithIsTestedAttribute() }));
-            Assert.Throws<PlatformNotSupportedException>(() => Activator.CreateInstance(typeof(MBRWithIsTestedAttribute), 0, null, null, CultureInfo.CurrentCulture, new object[] { new IsTestedAttribute(true) }));
-        }
-
-        [Fact]
-        static void TestingActivationAttributes1()
-        {
-            Activator.CreateInstance(typeof(ClassWithIsTestedAttribute), null, null);
-            Activator.CreateInstance(typeof(ClassWithIsTestedAttribute), null, new object[] { });
-
-        }
-    }
-}
