+++ src/System.Runtime/tests/System/ComponentModel/DefaultValueAttributeTests.cs
@@ -59,4 +59,44 @@ namespace System.ComponentModel.Tests
             }
         }
     }
+
+    public sealed class CustomDefaultValueAttribute : DefaultValueAttribute
+    {
+        public CustomDefaultValueAttribute(object value) : base(value) { }
+
+        public new void SetValue(object value) => base.SetValue(value);
+    }
+
+    public static class DefaultValueAttributeTestsNetStandard17
+    {
+        [Fact]
+        public static void SetValue()
+        {
+            var attr = new CustomDefaultValueAttribute(null);
+
+            attr.SetValue(true);
+            Assert.Equal(true, attr.Value);
+
+            attr.SetValue(false);
+            Assert.Equal(false, attr.Value);
+
+            attr.SetValue(12.8f);
+            Assert.Equal(12.8f, attr.Value);
+
+            attr.SetValue(12.8);
+            Assert.Equal(12.8, attr.Value);
+
+            attr.SetValue((byte)1);
+            Assert.Equal((byte)1, attr.Value);
+
+            attr.SetValue(28);
+            Assert.Equal(28, attr.Value);
+
+            attr.SetValue(TimeSpan.FromHours(1));
+            Assert.Equal(TimeSpan.FromHours(1), attr.Value);
+
+            attr.SetValue(null);
+            Assert.Null(attr.Value);
+        }
+    }
 }
