+++ src/System.Runtime/tests/System/HandleTests.cs
